Refine extra build configuration and test it properly
