+++ modules/@apostrophecms/asset/lib/webpack/utils.js
@@ -47,9 +47,12 @@ module.exports = {
     }
   },
 
-  // Export for testing
   formatRebundleConfig,
   verifyRebundleConfig,
+  verifyBundlesEntryPoints,
+  formatExtensionsOptions,
+  fillExtensionsOptions,
+  flattenBundles,
 
   transformRebundledFor,
 
@@ -76,40 +79,6 @@ module.exports = {
       verifiedBundles,
       rebundleModules
     };
-  },
-
-  async getBuildExtensions({
-    getMetadata, modulesToInstantiate, rebundleModulesConfig = {}
-  }) {
-    const modulesMeta = [];
-
-    for (const name of modulesToInstantiate) {
-      modulesMeta.push(await getMetadata(name));
-    }
-
-    const rebundleModules = formatRebundleConfig(rebundleModulesConfig);
-
-    const configsPerAlias = getModulesBuildConfigs(
-      modulesMeta,
-      rebundleModules
-    );
-
-    const configs = {};
-
-    for (const [ alias, config ] of Object.entries(configsPerAlias)) {
-      const {
-        extensions, extensionOptions, foundBundles
-      } = config;
-      const verifiedBundles = await verifyBundlesEntryPoints(foundBundles);
-      configs[alias] = {
-        extensions,
-        extensionOptions,
-        verifiedBundles,
-        rebundleModules
-      };
-    }
-
-    return configs;
   },
 
   fillExtraBundles (verifiedBundles = {}) {
@@ -291,71 +260,6 @@ function getModulesWebpackConfigs (modulesMeta, rebundleModules) {
   };
 };
 
-function getModulesBuildConfigs(modulesMeta, rebundleModules) {
-  const result = {};
-
-  // 1. for every "vendor alias" in the meta
-  for (const meta of modulesMeta) {
-    const { build, __meta } = meta;
-    const configsPerAlias = formatBuildConfigs({
-      chain: __meta.chain,
-      buildConfigs: build,
-      rebundleModules
-    });
-
-    for (const [ alias, configs ] of Object.entries(configsPerAlias)) {
-      if (!configs.length) {
-        continue;
-      }
-      if (!result[alias]) {
-        result[alias] = {
-          extensions: {},
-          extensionOptions: [],
-          bundles: {}
-        };
-      }
-
-      const extensionOptions = configs.reduce((acc, { extensionOptions = {} }) => {
-        return [
-          ...acc,
-          extensionOptions
-        ];
-      }, []);
-
-      result[alias].extensions = Object.assign(
-        result[alias].extensions,
-        reduce(configs, 'extensions')
-      );
-      result[alias].extensionOptions.push(...extensionOptions);
-      result[alias].bundles = Object.assign(
-        result[alias].bundles,
-        reduce(configs, 'bundles')
-      );
-    }
-  }
-
-  return Object.entries(result).reduce((acc, [ alias, config ]) => {
-    const {
-      extensions, extensionOptions, bundles
-    } = config;
-    const formattedOptions = formatExtensionsOptions(extensionOptions);
-    const { exts, options } = fillExtensionsOptions(extensions, formattedOptions);
-    acc[alias] = {
-      extensions: exts,
-      extensionOptions: options,
-      foundBundles: flattenBundles(bundles)
-    };
-    return acc;
-  }, {});
-
-  function reduce(list, prop) {
-    return list.reduce((acc, cur) => ({
-      ...acc,
-      ...cur[prop] || {}
-    }), {});
-  }
-};
-
 async function verifyBundlesEntryPoints (bundles) {
   const checkPathsPromises = bundles.map(async ({
     bundleName, modulePath, bundleRemapping
@@ -406,6 +310,7 @@ async function verifyBundlesEntryPoints (bundles) {
 
     return {
       bundleName,
+      modulePath,
       main,
       withIndex,
       ...jsFileExists && { jsPath },
@@ -447,7 +352,7 @@ async function verifyBundlesEntryPoints (bundles) {
   }
 
   const packedFilesByBundle = bundlesPathsWithIndex.reduce((acc, {
-    bundleName, jsPath, scssPath, main, withIndex
+    bundleName, modulePath, jsPath, scssPath, main, withIndex
   }) => {
     if (!jsPath && !scssPath) {
       return acc;
@@ -460,7 +365,8 @@ async function verifyBundlesEntryPoints (bundles) {
         // Boolean indicating if the "main" index.js is included
         // caused by a remapping. It's not yet used anywhere but
         // it's an useful information that we keep.
-        withIndex: !!withIndex || acc[bundleName]?.withIndex,
+        withIndex: !!withIndex || acc[bundleName]?.withIndex || false,
+        modulePath,
         js: [
           ...withIndex?.jsPath ? [ withIndex?.jsPath ] : [],
           ...acc[bundleName] ? acc[bundleName].js : [],
@@ -609,53 +515,6 @@ function formatConfigs (chain, webpackConfigs, rebundleModules) {
         }
       };
     }).filter((config) => config);
-}
-
-function formatBuildConfigs({
-  chain, buildConfigs, rebundleModules, result = {}
-}) {
-  Object.entries(buildConfigs)
-    .forEach(([ name, options ], i) => {
-
-      if (!options) {
-        return null;
-      }
-
-      for (const [ alias, config ] of Object.entries(options)) {
-        if (!Object.keys(config).length) {
-          continue;
-        }
-        if (!result[alias]) {
-          result[alias] = [];
-        }
-
-        const {
-          bundles = {},
-          extensions = {},
-          extensionOptions = {}
-        } = config;
-        const bundleNames = Object.keys(bundles);
-
-        // Remapping only for the explicitly defined by the module bundle
-        // configurations
-        const bundleRemapping = rebundleModules
-          .filter(entry => entry.name === chain[i].name);
-
-        result[alias].push({
-          name: chain[i].name,
-          extensions,
-          extensionOptions,
-          bundles: {
-            [name]: {
-              bundleNames,
-              modulePath: chain[i].dirname,
-              bundleRemapping
-            }
-          }
-        });
-      }
-    });
-  return result;
 }
 
 function flattenBundles (bundles) {
