+++ test/modules/article-page/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  extend: '@apostrophecms/piece-page-type',
+  init() { }
+};
