+++ test/modules/article-page/ui/src/index.js
@@ -1,0 +1,6 @@
+export default () => {
+  window.TESTBED = {
+    ...(window.TESTBED || {}),
+    articlePageIndex: true
+  };
+};
