+++ test/modules/article-page/ui/src/main.js
@@ -1,0 +1,6 @@
+export default () => {
+  window.TESTBED = {
+    ...(window.TESTBED || {}),
+    articlePageMain: true
+  };
+};
