+++ test/modules/article-page/ui/src/main.scss
@@ -1,0 +1,3 @@
+.test-article-page-main {
+  font-weight: 300;
+}
