+++ test/modules/article-widget/index.js
@@ -1,0 +1,9 @@
+module.exports = {
+  extend: '@apostrophecms/widget-type',
+  options: {
+    name: 'article',
+    label: 'Recent Articles',
+    icon: 'text-icon'
+  },
+  init() { }
+};
