+++ test/modules/article-widget/ui/src/carousel.js
@@ -1,0 +1,6 @@
+export default () => {
+  window.TESTBED = {
+    ...(window.TESTBED || {}),
+    articleWidgetCarousel: true
+  };
+};
