+++ test/modules/article-widget/ui/src/carousel.scss
@@ -1,0 +1,4 @@
+.test-carousel {
+  font-weight: bold;
+}
+
