+++ test/modules/article-widget/ui/src/topic.js
@@ -1,0 +1,6 @@
+export default () => {
+  window.TESTBED = {
+    ...(window.TESTBED || {}),
+    articleWidgetTopic: true
+  };
+};
