+++ test/modules/selected-article-widget/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+  extend: '@apostrophecms/widget-type',
+  options: {
+    label: 'Selected Articles'
+  },
+  init() { }
+};
