+++ test/modules/selected-article-widget/ui/src/tabs.js
@@ -1,0 +1,6 @@
+export default () => {
+  window.TESTBED = {
+    ...(window.TESTBED || {}),
+    selectedArticleWidgetTabs: true
+  };
+};
